Add Bevy selection focus and agent follow camera
