+++ .pm/tasks/task_b0db62ebc21b4f81815398a66c84f366.execution.md
@@ -1,0 +1,87 @@
+# task_b0db62ebc21b4f81815398a66c84f366 Execution Log
+
+- task_uid: task_b0db62ebc21b4f81815398a66c84f366
+- title: local launcher and game playtest on latest main
+- owner_role: tpm
+- worktree_hint: /home/scc/worktrees/oasis7-viewer-local-launch-playtest-main
+
+<!-- Append entries using:
+Example:
+  ## YYYY-MM-DD HH:MM:SS CST / role_name
+  - 完成内容: ...
+  - 遗留事项: ...
+  - Action: ...
+  - Validation Command: ...
+  - Expected Result: ...
+  - Actual Result: ...
+  - Blocker / Next Action: ...
+-->
+
+## 2026-06-07 12:00:30 CST / tpm
+- 完成内容: WORKFLOW BOOTSTRAP DECIDED
+- Repository State Impact: changes local runtime/build state by creating a dedicated latest-main task worktree, preparing/reusing launcher artifacts, and starting a local playtest stack for the user.
+- Isolation Decision: previous playtest worktree had accumulated task truth and generated artifacts, so it was not reused for latest-main validation; created `/home/scc/worktrees/oasis7-viewer-local-launch-playtest-main` from current local `main`.
+- Task Truth: owner role `tpm`; task `.pm/tasks/task_b0db62ebc21b4f81815398a66c84f366.yaml`; execution log `.pm/tasks/task_b0db62ebc21b4f81815398a66c84f366.execution.md`.
+- Routed Next Phase: `repo-owned-workflow-router` -> execution plus reachability verification. Direction is concrete: compile/start launcher and game from the current local `main` so the user can play the newest local code.
+- Required Writeback: `.pm` execution log mandatory; no PRD/project change expected unless startup exposes product-truth drift.
+- Subagent Slice Plan:
+  - role: `viewer_engineer`
+  - slice type: bounded launch-surface execution/verification evidence
+  - intended model configuration: workflow default subagent runtime
+  - actual dispatched model/reasoning: inherited/unverified; no subagent spawned in this turn, so TPM records the role contract and executes the local evidence path directly.
+  - context delivery mode: full-thread/full-history fork intended; fallback is this execution-log checklist.
+  - mandatory context checklist/packet: AGENTS.md workflow authority; `.agents/roles/viewer_engineer.md`; `scripts/run-producer-playtest.sh`; `scripts/run-game-test.sh`; task UID/worktree path; user intent: "用最新main".
+  - write scope: no source edits expected; runtime/build artifacts and `.pm` evidence only.
+  - return contract: latest-main commit id, playable URL, startup log path, and user-facing caveat if bundle mode is unavailable.
+  - formal sink / writeback surface: this execution log.
+  - integration owner: tpm.
+  - integration order: launcher/game startup first, reachability verification second.
+- Validation Command: `git -C /home/scc/worktrees/oasis7-viewer-local-launch-playtest-main rev-parse HEAD`
+- Actual Result: `cbc50ec01c6ff52ea88f19cd5d496d1b5085a50d`, matching the current local `main` at task creation time.
+- Blocker / Next Action: none; start the latest-main local playtest stack.
+
+## 2026-06-07 12:42:00 CST / tpm
+- 完成内容: execution implementation started for left-rail selection camera focus.
+- User Intent: user asked whether selecting an agent should make the Bevy pixel renderer follow/focus, then asked to implement it directly.
+- Scope Decision: keep the change local to `crates/pixel_world_bridge` by treating selection changes as a camera-focus request without changing frontend selection payload shape.
+- Implementation Plan:
+  - detect selection delta during `apply_external_render_snapshot`
+  - persist a pending focus target in Bevy runtime state
+  - apply focus during render once window dimensions are known
+  - preserve existing drag/wheel manual camera semantics outside explicit selection-focus transitions
+- Validation Plan: targeted `pixel_world_bridge` tests for selection-only camera behavior plus existing selection render regressions; then rerun local playtest stack if tests pass.
+
+## 2026-06-07 14:35:00 CST / tpm
+- 完成内容: implemented Bevy selection focus plus continuous selected-agent follow.
+- Changed Paths: `crates/pixel_world_bridge/src/lib.rs`; `crates/pixel_world_bridge/src/render.rs`; regenerated `crates/oasis7_viewer/pixel-world-bridge/pixel_world_bridge_bindgen_bg.wasm`.
+- Behavior Change:
+  - selecting an `agent` or `location` now recenters the Bevy pixel-world camera on the selected entity
+  - selecting an `agent` additionally enables follow mode so later position updates keep the camera centered on that same agent
+  - manual drag or wheel input clears follow mode so user camera control wins immediately
+- Validation Command: `env -u RUSTC_WRAPPER cargo test -p pixel_world_bridge -- --nocapture`
+- Actual Result: pass, 19 tests passed / 0 failed, including new selection/follow tests and existing selection visibility regressions.
+- Runtime Verification: restarted local stack on latest local `main`; latest ready run is `startup-20260607-143112` with playable URL `http://127.0.0.1:4173/?ws=ws://127.0.0.1:5011&test_api=1&locale=zh`.
+- Next Action: close task, record pre-PR local review packet with tool limitation caveat, then commit and prepare PR.
+
+## 2026-06-07 14:36:00 CST / tpm
+- Review Trigger: pre-PR local role review
+- Review Scope: `crates/pixel_world_bridge/src/lib.rs`; `crates/pixel_world_bridge/src/render.rs`; regenerated `crates/oasis7_viewer/pixel-world-bridge/pixel_world_bridge_bindgen_bg.wasm`
+- Review Roles: `viewer_engineer`, `qa_engineer`
+- Review Question: does the diff correctly implement selection focus + continuous selected-agent follow without regressing visibility, manual camera override, or latest-main local playability?
+- Evidence Available: `env -u RUSTC_WRAPPER cargo test -p pixel_world_bridge -- --nocapture`; latest-main local playtest run `startup-20260607-143112`; task execution log evidence above
+- Expected Return Contract: `findings | no_findings | residual_risk`
+- Formal Sink: this execution log
+- Review Dispatch Limitation: current connector policy in this thread does not permit spawning subagents without explicit user delegation permission, so TPM records the required role review packet and uses repo-truth + fresh local verification evidence as fallback instead of fabricating subagent output.
+- Pre-PR Local Role Review: passed
+- Task UID: task_b0db62ebc21b4f81815398a66c84f366
+- Source Worktree: /home/scc/worktrees/oasis7-viewer-local-launch-playtest-main
+- Source Branch: task/viewer-local-launch-playtest-main
+- Source Head: cbc50ec01c6ff52ea88f19cd5d496d1b5085a50d plus uncommitted task diff under review
+- Comparison Ref: main
+- Reviewed Changed Paths: crates/pixel_world_bridge/src/lib.rs; crates/pixel_world_bridge/src/render.rs; crates/oasis7_viewer/pixel-world-bridge/pixel_world_bridge_bindgen_bg.wasm; .pm task truth files
+- Role Selection Basis: viewer-owned player-facing world camera behavior and QA-owned behavioral verification were both directly touched; skipped liveops/community and producer/system roles because no external messaging or product-policy change landed
+- Review Roles: viewer_engineer, qa_engineer
+- Review Evidence: viewer_engineer fallback evidence = code-path review over selection/focus/follow state transitions; qa_engineer fallback evidence = fresh `cargo test -p pixel_world_bridge` pass plus latest-main stack restart evidence
+- Review Findings Disposition: no_findings
+- Finding Disposition Evidence: no valid findings emerged beyond the documented subagent-dispatch tool limitation
+- Residual Risk: browser-level manual validation of long-running agent motion follow is still interactive/runtime-dependent; automated coverage currently proves selection-change and follow-state propagation but not a full end-to-end browser camera trace.
